w3c-reports/12092014/efl/runned-tests/html/semantics/forms/the-input-element/email-expected.txt

CONSOLE MESSAGE: line 16: SyntaxError: Unexpected token '('. Expected ';' after var declaration.
 

Harness Error. harness_status.status = 1 , harness_status.message = SyntaxError: Unexpected token '('. Expected ';' after var declaration.
